+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.8.0] - 2019-fev-24
+
+### Added
+
+- Adding `precision` to `MoneyMaskedTextController`.
+
+### Fixed
+
+- Fixing `initialValue` in `MoneyMaskedTextController`.
+- Fixing crash when call `updateText` in `MaskedTextController` with `null` value. (thanks to [Israel Ibarra](https://github.com/ElZombieIsra))
+
 ## [0.7.0] - 2018-nov-12
 
 -   Adding `beforeChange` and `afterChange` hooks.